Remove EDH fitness leaderboard api methods & tests
